# specs/openapi/gits-kno-api.openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "GITS Knowledge Engineering API",
    "version": "1.1.0",
    "description": "GITS 知识工程平台 API — V1.1"
  },
  "servers": [
    {
      "url": "/api/v1"
    }
  ],
...
  "paths": {
...
    "/engagement/journey/start": {
      "post": {
        "operationId": "startJourney",
        "tags": [
          "Engagement Journey"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JourneyStartRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyStartResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/engagement/journey/{journeyId}/previsit": {
      "post": {
        "operationId": "executePrevisit",
        "tags": [
          "Engagement Journey"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/journeyId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PrevisitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrevisitExecutionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/engagement/journey/{journeyId}/postvisit": {
      "post": {
        "operationId": "executePostvisit",
        "tags": [
          "Engagement Journey"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/journeyId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PostvisitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PostvisitExecutionResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/engagement/journey/{journeyId}/new-evidence": {
      "post": {
        "operationId": "processNewEvidence",
        "tags": [
          "Engagement Journey"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/journeyId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewEvidenceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NewEvidenceResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/engagement/journey/{journeyId}/complete": {
      "post": {
        "operationId": "completeJourney",
        "tags": [
          "Engagement Journey"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/journeyId"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JourneyCompleteResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
    "/engagement/journey/outreach-script": {
      "post": {
        "operationId": "generateOutreachScript",
        "tags": [
          "Engagement Journey"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutreachScriptRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutreachScriptResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/engagement/journey/meeting-script": {
      "post": {
        "operationId": "generateMeetingScript",
        "tags": [
          "Engagement Journey"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MeetingScriptRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeetingScriptResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/engagement/journey/outreach-scripts": {
      "get": {
        "operationId": "listOutreachScripts",
        "tags": [
          "Engagement Journey"
        ],
        "parameters": [
          {
            "name": "customerId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OutreachScript"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/engagement/journey/meeting-scripts": {
      "get": {
        "operationId": "listMeetingScripts",
        "tags": [
          "Engagement Journey"
        ],
        "parameters": [
          {
            "name": "customerId",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MeetingScript"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/engagement/journey/{journeyId}/prepare-previsit": {
      "post": {
        "operationId": "preparePrevisit",
        "tags": [
          "Engagement Journey"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/journeyId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreparedPrevisitRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreparedPrevisitResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "journeyId": {
        "name": "journeyId",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      },
      "Conflict": {
        "description": "业务冲突",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        }
      }
    },
    "schemas": {
...
      "Problem": {
        "type": "object",
        "required": [
          "status",
          "error",
          "message",
          "path",
          "timestamp"
        ],
        "properties": {
          "status": {
            "type": "integer"
          },
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "JourneyStartRequest": {
        "type": "object",
        "required": [
          "customerId"
        ],
        "properties": {
          "customerId": {
            "type": "string"
          },
          "journeyType": {
            "type": "string"
          },
          "rmId": {
            "type": "string"
          },
          "context": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "JourneyStartResponse": {
        "type": "object",
        "required": [
          "journeyId",
          "customerId",
          "phase",
          "startedAt"
        ],
        "properties": {
          "journeyId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "phase": {
            "type": "string"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time"
          },
          "previsitReady": {
            "type": "boolean"
          }
        }
      },
      "PrevisitRequest": {
        "type": "object",
        "properties": {
          "includeQuickBattleCard": {
            "type": "boolean"
          },
          "focusAreas": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "PrevisitExecutionResponse": {
        "type": "object",
        "required": [
          "journeyId",
          "reportGenerated"
        ],
        "properties": {
          "journeyId": {
            "type": "string"
          },
          "reportGenerated": {
            "type": "boolean"
          },
          "reportContent": {
            "$ref": "#/components/schemas/PrevisitReportContent"
          },
          "quickBattleCard": {
            "$ref": "#/components/schemas/QuickBattleCard"
          },
          "kycGapsIdentified": {
            "type": "integer"
          },
          "signalsDetected": {
            "type": "integer"
          },
          "assemblyTrace": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssemblyTraceStep"
            }
          },
          "skillReportTitle": {
            "type": "string"
          },
          "skillExecutiveSummary": {
            "type": "string"
          },
          "skillSections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillReportSection"
            }
          }
        }
      },
      "PostvisitRequest": {
        "type": "object",
        "properties": {
          "interactionId": {
            "type": "string"
          },
          "transcript": {
            "type": "string"
          }
        }
      },
      "PostvisitExecutionResponse": {
        "type": "object",
        "required": [
          "journeyId",
          "analysisGenerated"
        ],
        "properties": {
          "journeyId": {
            "type": "string"
          },
          "analysisGenerated": {
            "type": "boolean"
          },
          "analysisContent": {
            "$ref": "#/components/schemas/PostvisitAnalysisContent"
          },
          "claimsExtracted": {
            "type": "integer"
          },
          "commitmentsIdentified": {
            "type": "integer"
          },
          "opportunitiesDetected": {
            "type": "integer"
          }
        }
      },
      "NewEvidenceRequest": {
        "type": "object",
        "required": [
          "evidenceType",
          "content"
        ],
        "properties": {
          "evidenceType": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "confidence": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "NewEvidenceResponse": {
        "type": "object",
        "required": [
          "evidenceId",
          "journeyId"
        ],
        "properties": {
          "evidenceId": {
            "type": "string"
          },
          "journeyId": {
            "type": "string"
          },
          "reconciliationTriggered": {
            "type": "boolean"
          },
          "conflictingClaims": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "JourneyCompleteResponse": {
        "type": "object",
        "required": [
          "journeyId",
          "completedAt"
        ],
        "properties": {
          "journeyId": {
            "type": "string"
          },
          "completedAt": {
            "type": "string",
            "format": "date-time"
          },
          "summary": {
            "type": "object",
            "properties": {
              "totalClaims": {
                "type": "integer"
              },
              "totalCommitments": {
                "type": "integer"
              },
              "totalOpportunities": {
                "type": "integer"
              },
              "kycGapReduction": {
                "type": "integer"
              }
            }
          }
        }
      },
      "OutreachScriptRequest": {
        "type": "object",
        "required": [
          "customerId"
        ],
        "properties": {
          "customerId": {
            "type": "string"
          },
          "channel": {
            "type": "string"
          },
          "purpose": {
            "type": "string"
          },
          "tone": {
            "type": "string"
          }
        }
      },
      "OutreachScriptResponse": {
        "type": "object",
        "required": [
          "scriptId",
          "customerId"
        ],
        "properties": {
          "scriptId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "scriptContent": {
            "type": "string"
          },
          "channel": {
            "type": "string"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MeetingScriptRequest": {
        "type": "object",
        "required": [
          "customerId"
        ],
        "properties": {
          "customerId": {
            "type": "string"
          },
          "meetingType": {
            "type": "string"
          },
          "objectives": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "MeetingScriptResponse": {
        "type": "object",
        "required": [
          "scriptId",
          "customerId"
        ],
        "properties": {
          "scriptId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "scriptContent": {
            "type": "string"
          },
          "meetingType": {
            "type": "string"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PrevisitReportContent": {
        "type": "object",
        "properties": {
          "customerOverview": {
            "type": "string"
          },
          "kycGapSummary": {
            "type": "string"
          },
          "opportunityHighlights": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "riskIndicators": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "recommendedActions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "previousInteractionSummary": {
            "type": "string"
          }
        }
      },
      "QuickBattleCard": {
        "type": "object",
        "properties": {
          "customerName": {
            "type": "string"
          },
          "keyFacts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "openItems": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "suggestedTopics": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "riskAlerts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "OutreachScript": {
        "type": "object",
        "required": [
          "scriptId",
          "customerId",
          "channel"
        ],
        "properties": {
          "scriptId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "channel": {
            "type": "string"
          },
          "purpose": {
            "type": "string"
          },
          "scriptContent": {
            "type": "string"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MeetingScript": {
        "type": "object",
        "required": [
          "scriptId",
          "customerId"
        ],
        "properties": {
          "scriptId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "meetingType": {
            "type": "string"
          },
          "scriptContent": {
            "type": "string"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PreparedPrevisitRequest": {
        "type": "object",
        "required": [
          "customerId",
          "operatingCaseId"
        ],
        "properties": {
          "customerId": {
            "type": "string"
          },
          "rmId": {
            "type": "string"
          },
          "operatingCaseId": {
            "type": "string"
          },
          "journeyId": {
            "type": "string"
          },
          "channel": {
            "type": "string"
          },
          "visitObjective": {
            "type": "string"
          }
        }
      },
      "PreparedPrevisitResponse": {
        "type": "object",
        "properties": {
          "outreachScript": {
            "$ref": "#/components/schemas/OutreachScript"
          },
          "meetingScript": {
            "$ref": "#/components/schemas/MeetingScript"
          },
          "previsitReport": {
            "$ref": "#/components/schemas/PrevisitReportContent"
          },
          "battleCard": {
            "$ref": "#/components/schemas/QuickBattleCard"
          },
          "supplyChainMarkdown": {
            "type": "string"
          },
          "assemblyTrace": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssemblyTraceStep"
            }
          },
          "skillReportTitle": {
            "type": "string"
          },
          "skillExecutiveSummary": {
            "type": "string"
          },
          "skillSections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillReportSection"
            }
          }
        }
      },
      "AssemblyTraceStep": {
        "type": "object",
        "required": [
          "phase",
          "status",
          "message"
        ],
        "properties": {
          "phase": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "kiId": {
            "type": "string"
          }
        }
      },
...
    }
  }
}